Improve numerical accuracy of DRF calc at larger distances. Previous implementation used floats to perform solid-angle calculation; this would lead to numerical accuracy issues possibly starting around 12 meters! Converted to using doubles, which look to be good up to around 200 km.  I tried a few other forms of the equations, but they appeared to be less numerically accurate.

+++ src/FluxTool.cpp
@@ -862,10 +862,10 @@ void FluxToolWidget::refreshPeakTable()
   
   m_msg->setText( "" );
   
-  float distance = static_cast<float>( 1.0*PhysicalUnits::meter );
+  double distance = 1.0*PhysicalUnits::meter;
   try
   {
-    distance = static_cast<float>( PhysicalUnits::stringToDistance( m_distance->text().toUTF8() ) );
+    distance = PhysicalUnits::stringToDistance( m_distance->text().toUTF8() );
   }catch(...)
   {
     m_msg->setText( "Invalid Distance" );
